add pricingController
